JavascriptとViewの修正
