+++ app/controllers/flower_mountains_controller.rb
@@ -41,6 +41,10 @@ class FlowerMountainsController < ApplicationController
   def show
     @flower_mountain = FlowerMountain.includes(:flower, :mountain).find(params[:id])
     @posts = @flower_mountain.posts.recent.includes(:user).page(params[:page]).per(5)
+    # 関連する花山の取得
+    @related_flower_mountains = FlowerMountain.where(flower_id: @flower_mountain.flower_id)
+                                              .where.not(id: @flower_mountain.id)
+                                              .to_a  # ← 結果がnilにならない！
     
     # ログイン済みの場合、お気に入り状態を確認
     if user_signed_in?
