+++ app/views/flower_mountains/show.html.erb
@@ -45,17 +45,19 @@
         
         <!-- お気に入りボタン -->
         <% if user_signed_in? %>
-          <%= link_to favorite_mountain_flower_mountain_path(@flower_mountain.mountain, @flower_mountain), 
-                      method: :post, 
-                      class: "favorite-button btn #{'btn-warning' if @is_favorited} #{'btn-outline-warning' unless @is_favorited}", 
-                      data: { remote: true } do %>
-            <i class="fas fa-star"></i>
+          <%= link_to toggle_favorites_path(flower_mountain_id: @flower_mountain.id), # ルーティングで定義したtoggle_favorites_pathを使用
+                      class: "favorite-button btn #{current_user.favorites.exists?(flower_mountain: @flower_mountain) ? 'btn-warning' : 'btn-outline-warning'}",
+                      data: {
+                        # data-method や data-remote は削除
+                        # initFavoriteButtons が href と POST を使うため、これらは不要
+                      } do %>
+            <i class="bi <%= current_user.favorites.exists?(flower_mountain: @flower_mountain) ? 'bi-star-fill' : 'bi-star' %>"></i>
             <span class="favorite-count"><%= @favorites_count %></span>
           <% end %>
         <% else %>
           <button class="btn btn-outline-warning" disabled>
-            <i class="far fa-star"></i>
-            <%= @favorites_count %>
+          <i class="far fa-star"></i>
+          <%= @favorites_count %>
           </button>
         <% end %>
       </div>
@@ -168,7 +170,7 @@
         <div class="col-md-4 col-lg-3 mb-4">
           <div class="card border-0 shadow-sm hover-card h-100">
             <%= link_to post_path(post) do %>
-              <%= image_tag post.image_url.presence || "/placeholder.svg?height=200&width=200", 
+              <%= image_tag post.image_url.url.presence || "/placeholder.svg?height=200&width=200", 
                           class: "card-img-top", 
                           style: "height: 200px; object-fit: cover;" %>
             <% end %>
@@ -206,8 +208,8 @@
       <% @related_flower_mountains.each do |fm| %>
         <div class="col-md-4 mb-4">
           <div class="card border-0 shadow-sm hover-card h-100">
-            <%= link_to mountain_flower_mountain_path(fm.mountain, fm), class: "text-decoration-none" do %>
-              <%= image_tag fm.mountain.image_url.presence || "/placeholder.svg?height=200&width=300&text=#{fm.mountain.name}", 
+            <%= link_to flower_mountain_path(fm.mountain, fm), class: "text-decoration-none" do %>
+              <%= image_tag fm.mountain.image_url.url.presence || "/placeholder.svg?height=200&width=300&text=#{fm.mountain.name}", 
                           class: "card-img-top", 
                           style: "height: 200px; object-fit: cover;" %>
               <div class="card-body">
