+++ app/views/flowers/show.html.erb
@@ -7,7 +7,7 @@
       <div class="card border-0 shadow-lg overflow-hidden">
         <div class="row g-0">
           <div class="col-md-6">
-            <%= image_tag @flower.image_url.presence || "/placeholder.svg?height=400&width=600&text=#{@flower.name}", 
+            <%= image_tag @flower.image_url.url.presence || "/placeholder.svg?height=400&width=600&text=#{@flower.name}", 
                          class: "img-fluid w-100", 
                          style: "height: 400px; object-fit: cover;" %>
           </div>
@@ -30,14 +30,14 @@
               <div class="d-flex gap-3 mb-4">
                 <!-- いいねボタン -->
                 <% if user_signed_in? %>
-                  <% if @is_favorited %>
-                    <%= link_to favorite_path(@flower), method: :delete, class: "btn btn-warning" do %>
-                      <i class="bi bi-star-fill me-2"></i> お気に入り解除
-                    <% end %>
-                  <% else %>
-                    <%= link_to favorites_path(flower_mountain_id: @flower.flower_mountains.first.id), method: :post, class: "btn btn-outline-warning" do %>
-                      <i class="bi bi-star me-2"></i> お気に入り追加
-                    <% end %>
+                  <%= link_to toggle_favorites_path(flower_mountain_id: @flower_mountains.first.id), # flower_mountainsが複数ある場合は適切に選択
+                                class: "favorite-button btn btn-sm #{current_user.favorites.exists?(flower_mountain: @flower_mountains.first) ? 'btn-warning' : 'btn-outline-warning'}",
+                                data: {
+                                # JavaScriptの initFavoriteButtons が href と POST を使うため、data-method/data-turbo-method は不要
+                                # data-flower-mountain-id: @flower_mountains.first.id # JSで必要なら追加
+                                } do %>
+                    <i class="bi <%= current_user.favorites.exists?(flower_mountain: @flower_mountains.first) ? 'bi-star-fill' : 'bi-star' %> me-2"></i>
+                    <%= current_user.favorites.exists?(flower_mountain: @flower_mountains.first) ? 'お気に入り解除' : 'お気に入りに追加' %>
                   <% end %>
                   <span class="ms-2 text-muted"><%= @favorites_count %>人がお気に入り</span>
                 <% else %>
@@ -201,7 +201,7 @@
           <% @recent_posts.each do |post| %>
             <div class="col-6 col-md-4 col-lg-3 mb-4">
               <div class="card border-0 shadow-sm hover-card">
-                <%= image_tag post.image_url.presence || "/placeholder.svg?height=200&width=200", 
+                <%= image_tag post.image_url.url.presence || "/placeholder.svg?height=200&width=200", 
                              class: "card-img-top", 
                              style: "height: 200px; object-fit: cover;" %>
                 <div class="card-body p-3">
