+++ app/views/mountains/index.html.erb
@@ -188,8 +188,8 @@
       <% if @mountains.respond_to?(:current_page) %>
         <div class="row">
           <div class="col-12">
-            <!--div class="d-flex justify-content-center"-->
-              <%#= paginate @mountains %>
+            <div class="d-flex justify-content-center">
+              <%= paginate @mountains %>
             </div>
           </div>
         </div>
@@ -205,15 +205,15 @@
       </div>
     </div>
   </div>
+</div>
 
-  <!-- 管理者用ボタン 現状はなしのち追加：? && current_user.admin?-->
-  <% if defined?(user_signed_in?) && user_signed_in? %>
+  <!-- 管理者用ボタン 現状はなしのち追加：? && current_user.admin?
+  <%# if defined?(user_signed_in?) && user_signed_in? %>
     <div class="position-fixed bottom-0 end-0 m-4">
-      <%= link_to new_mountain_path, class: "btn btn-success rounded-circle shadow-lg", 
+      <%#= link_to new_mountain_path, class: "btn btn-success rounded-circle shadow-lg", 
                 style: "width: 60px; height: 60px;", 
                 data: { bs_toggle: "tooltip", bs_title: "新しい山を追加" } do %>
         <i class="bi bi-plus-lg"></i>
-      <% end %>
+      <%# end %>
     </div>
-  <% end %>
-</div>
+  <%# end %>-->
