--- app/views/shared/_header.html.erb
@@ -1,73 +1,0 @@
-<header class="navbar navbar-expand-lg navbar-light bg-white shadow-sm sticky-top">
-  <div class="container">
-    <a class="navbar-brand d-flex align-items-center" href="<%= root_path %>">
-      <i class="bi bi-flower1 text-success me-2"></i>
-      <span class="fw-bold text-gradient">花山リマインダー</span>
-    </a>
-    
-    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav">
-      <span class="navbar-toggler-icon"></span>
-    </button>
-    
-    <div class="collapse navbar-collapse" id="navbarNav">
-      <ul class="navbar-nav ms-auto align-items-center">
-        <li class="nav-item">
-          <a class="nav-link" href="<%= flower_mountains_path %>">
-            <i class="bi bi-map me-1"></i>花登山マップ
-          </a>
-        </li>
-        
-        <li class="nav-item">
-          <a class="nav-link" href="<%= posts_path %>">
-            <i class="bi bi-camera me-1"></i>写真ランキング
-          </a>
-        </li>
-        
-        <% if user_signed_in? %>
-          <li class="nav-item">
-            <a class="nav-link" href="<%= favorites_path %>">
-              <i class="bi bi-heart me-1"></i>お気に入り
-            </a>
-          </li>
-          
-          <li class="nav-item">
-            <a class="nav-link" href="<%= notifications_path %>">
-              <i class="bi bi-bell me-1"></i>通知設定
-            </a>
-          </li>
-          
-          <li class="nav-item">
-            <%= link_to new_post_path, class: "btn btn-pink btn-sm ms-2" do %>
-              <i class="bi bi-upload me-1"></i>投稿する
-            <% end %>
-          </li>
-          
-          <li class="nav-item dropdown ms-2">
-            <a class="nav-link dropdown-toggle" href="#" id="userDropdown" role="button" data-bs-toggle="dropdown">
-              <i class="bi bi-person-circle me-1"></i><%= current_user.nickname || "マイページ" %>
-            </a>
-            <ul class="dropdown-menu dropdown-menu-end">
-              <li>
-                <%= link_to "プロフィール", user_path(current_user), class: "dropdown-item" %>
-              </li>
-              <li>
-                <%= link_to "設定", edit_user_registration_path, class: "dropdown-item" %>
-              </li>
-              <li><hr class="dropdown-divider"></li>
-              <li>
-                <%= link_to "ログアウト", destroy_user_session_path, data: { turbo_method: :delete }, class: "dropdown-item" %>
-              </li>
-            </ul>
-          </li>
-        <% else %>
-          <li class="nav-item">
-            <%= link_to "ログイン", new_user_session_path, class: "btn btn-outline-success btn-sm ms-2" %>
-          </li>
-          <li class="nav-item">
-            <%= link_to "新規登録", new_user_registration_path, class: "btn btn-success btn-sm ms-2" %>
-          </li>
-        <% end %>
-      </ul>
-    </div>
-  </div>
-</header>
